+++ apps/dev/src/main.tsx
@@ -14,6 +14,7 @@ const gameRegistry: GameRegistry = {
     import("@jgengine/shell/demo/pointerDemo").then((module) => module.pointerDemoGame),
   "environment-showcase": () =>
     import("@jgengine/shell/demo/environmentShowcase").then((module) => module.environmentShowcaseGame),
+  "extraction-map": () => import("@jgengine/shell/demo/mapDemo").then((module) => module.mapDemoGame),
   "world-of-warcraft": () => import("@dogfood/world-of-warcraft").then((module) => module.wowGame),
   "asset-showcase": () =>
     import("@dogfood/asset-showcase").then((module) => module.assetShowcaseGame),